Defect fixes for PPC64 (big/little endian) platform merged from users/dmorash/Platform-8.3.
This includes:

1) Correctly sized DCE/RPC NDR types for PPC64 platforms; DCE/RPC 
  supports 2 and 4 byte integers.  Where the compiler supports C99 
  stdint types are used. With gcc this requires compiling with the
  -std=c99 or -std=gnu99 flag.

2) Corrected an issue where the rpc assoc group id (required to be
  4 bytes) was being set from an rpc client handle.  Client handles
  are meant to be opaque handles but are implemented as void *.

3) Fixed issues in the idl_compiler client stub generation code. 
  The generated idl type vector which describes how to marshall rpc
  call parameters was using long types for 4 byte data types. Since
  it only used the first four bytes of the long, the idl type vector
  was essentially empty and parameters for rpc calls would fail
  to marshall.

4) Corrects various LDAP related segfaults caused when getting the 
  maxbufsize option, and the result of LwLdapCountEntries. These 
  were writing larger types into local variables resulting in 
  corruption to other local variables.

5) Corrected dce rpc crash on PPC64 by synchronizing the definitions of
  the rpc_addr_p_t and rpc_addr_ip_t structs.

6) MakeKit module modifications to support PPC64.  This includes PPC32 
  but those are untested. 

7) Support for linux-ppc64 build of OpenSSL by setting appropriate 
   OpenSSL configure host_string value. e.g. linux-ppc64

8) Defect fix for gRegEnumValueRespSpec; the processing of the 
  CMD_TERMINATION subelement of the 
  LWMSG_ATTR_LENGTH_MEMBER(gRegEnumValueRespSpec, cName) was ultimately 
  causing a segfault. This was caused by the incorrect ordering of 
  elements in the spec, specifically cName value which indicates the 
  length of pName, should precede pName, and then be followed by 
  LWMSG_ATTR_LENGTH_MEMBER(..., cName).

9) Various compilation fixes (removal of unused variables).

10) Disable compiler optimizations when building cyrus-sasl on powerpc 
   otherwise duplicate symbols prevent linking.

11) Additionally fixed a segfault in threadpool-common WorkLoop related 
   to locking a pthread mutex.  This was only occuring when compiled with 
   optimization enabled (-O2) which enables strict-aliasing.  Fixed this 
   issue by removing the local variable pThreads which was an alias for
   pThread->pThreads, and rewriting affected lines to use pThread->pThreads. 


git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/branches/Platform-8.3/src/linux@69628 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ dcerpc/ncklib/cnsasgsm.c
@@ -872,7 +872,7 @@ INTERNAL unsigned32     rundown_handles_action_rtn
      */  
     if (assoc_grp->grp_liveness_mntr != NULL)
     {
-        (*assoc_grp->grp_liveness_mntr) ((rpc_client_handle_t) assoc_grp->grp_id.all);
+        (*assoc_grp->grp_liveness_mntr) ((rpc_client_handle_t) ((size_t)assoc_grp->grp_id.all));
     }
 
     sm_p->cur_state = RPC_C_ASSOC_GRP_CLOSED;
